scripts/fix_notebook_search_parameters.py
#!/usr/bin/env python3
"""
Fix for notebook search parameter issues.
Shows correct way to use search parameters with PatentSearcher.
"""

def parse_date_range(date_range_str):
    """Parse SEARCH_YEARS string into start_date and end_date."""
    if " to " in date_range_str:
        start_date, end_date = date_range_str.split(" to ")
        return start_date.strip(), end_date.strip()
    else:
        raise ValueError(f"Invalid date range format: '{date_range_str}'. Expected format: 'YYYY-MM-DD to YYYY-MM-DD'")

def get_search_config_limits():
    """Get search limits from configuration."""
    try:
        import yaml
        from pathlib import Path
        
        config_path = Path(__file__).parent.parent / 'config' / 'search_patterns_config.yaml'
        with open(config_path, 'r') as f:
            config = yaml.safe_load(f)
        
        return config['global_settings']['max_results']
    except Exception:
        # Fallback limits
        return {
            'default': 1000,
            'comprehensive': 5000,
            'focused': 500,
            'validation': 100
        }

# CORRECTED NOTEBOOK CODE EXAMPLE:
print("📋 CORRECTED SEARCH PARAMETERS:")
print("=" * 50)

# Define search parameters
SEARCH_TECHNOLOGY_AREAS = ["rare_earth_elements"]
SEARCH_YEARS = "2010-01-01 to 2024-12-31"  # Now will be used correctly
MAX_RESULTS_MODE = "comprehensive"  # Use config-defined limits

# Parse date range correctly
start_date, end_date = parse_date_range(SEARCH_YEARS)
print(f"📅 Parsed date range: {start_date} to {end_date}")

# Get config limits
config_limits = get_search_config_limits()
print(f"📊 Available result limits: {config_limits}")
print(f"📊 Selected limit mode: {MAX_RESULTS_MODE} ({config_limits[MAX_RESULTS_MODE]} results)")

print("\n📝 CORRECTED SEARCH CALL:")
print("""
# Execute technology-specific search (CORRECTED VERSION)
search_results = patent_searcher.execute_technology_specific_search(
    technology_areas=SEARCH_TECHNOLOGY_AREAS,
    start_date=start_date,  # Use parsed date
    end_date=end_date,      # Use parsed date
    focused_search=False    # Use comprehensive search for more results
)
""")

print("✅ This fixes the date range and result limit confusion!")
print("✅ Now SEARCH_YEARS parameter will actually be used!")
print("✅ Search method will respect the full date range!")