requirements.txt
# Patent Intelligence Platform - Production Dependencies
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0

# Database & API Integration
sqlalchemy>=2.0.0
requests>=2.28.0
google-cloud-bigquery>=3.4.0

# Visualization & Dashboards
plotly>=5.15.0
folium>=0.14.0
openpyxl>=3.1.0

# Geographic & Country Data
pycountry>=22.0.0
pycountry-convert>=0.7.2

# Development & Testing
jupyter>=1.0.0
pytest>=7.0.0
pytest-cov>=4.0.0

# Network Analysis (for citation networks)
networkx>=3.0

# Optional: EPO PATSTAT Integration (TIP Environment)
# epo-tipdata-patstat  # Available only in EPO TIP environment

# Production Web Framework (if needed)
flask>=2.3.0
gunicorn>=20.1.0

# Security & Authentication
cryptography>=40.0.0
python-dotenv>=1.0.0

# Performance & Monitoring
psutil>=5.9.0
memory-profiler>=0.60.0

# Data Export & Reporting
matplotlib>=3.6.0
seaborn>=0.12.0